requirements.txt
networkx==3.4.2
numpy==2.2.3
scipy==1.14.1
torch_geometric==2.6.1
tqdm==4.66.5

